ChatBridgeReforged_MC/requirements.txt
trio
mcdreforged
pycryptodome